Added nstp_config role
This role generates a server configuration script (PowerShell) for the Internet DNS/NTP (NSTP) server.  This script is applied after the server has been deployed.  It will configure the DNS and NTP services on the server, including adding DNS zone information, as well as populate the database with various records.